* master:
  Update pot files using current code and production database.
  Update translations from transifex.
  MBS-9827: Redirect merged entities (#742)
  Add tests for Niconico URL handler (MBS-9944)
  MBS-9944 (2): Handle Niconico Video URLs
  MBS-9944 (1): Display Niconico URL in the sidebar
  Add Bandcamp URL cleanup test for (MBS-9831)
  MBS-9831: Handle bandcamp URL for work lyrics
  Add component for inline list of users (#770)
  MBS-8178: Create report for sequential events not in series (#767)
  MBS-6232: Report for releases w/o CAA type (#744)
  MBS-9872: Fix change to bracketedWithSpace (#769)
  MBS-9402: Fix multilang display (#768)
  MBS-7192: Show work types in inline search (#728)
  Add new column to ReleaseLabelSameArtist report
  Improve Label Report from #760
  Fix a typo in HACKING doc
  Add search feature to PlacesWithoutCoordinates report
  Fix images file permission
  MBS-9930: Disable CAA image on homepage for Chrome (#759)
  MBS-9907: Report for places without coordinates
  Load Area entities into Place Report
  Add JS xgettext replacement
  Update pot files using current code and production database.
  Switch to project-wide Babel config
  MBS-9934: Update blog and JIRA links to metabrainz.org (#761)
  Text changes for MBS-3685
  Bump Flow to 0.87.0
  MBS-3685: Report for releases where label name = artist name
  Add test cfor Tipeee language subdomain (MBS-9905)
  Clean some regular expressions in URLCleanup
  Add test for Mixcloud URL handler (MBS-9277)
  MBS-9277: Display Mixcloud URLs in the sidebar
  MBS-9277: Add autoselect/cleanup for Mixcloud URLs
  MBS-8311: Inform about release annotation license (#756)
  Remove unused import from ArtistResults
  MBS-9905: Language subdomain support for Tipeee
  MBS-9765: Do not display event dates twice in search results (#741)
  Small tweaks to PR #752 (MBS-9900) (#755)
  Drop WWW::Mechanize::tick from test
  Update to IANA TimeZone Database 2018e
  Allow generating snapshot for any cpanfile
  Remove hard-coded duplicate UTC timezone option
  Rewrite Preferences page to React/JSX
  Rewrite user/account page layout to React/JSX
  Rewrite user/account tabs to React/JSX
  Allow adding helper buttons beside FormRowSelect
  Allow handling change event for FormRowCheckbox
  Complete editor preferences serialization to JSON
  Add editor's deleted flag to JSON/React
  Fix typo in (yet unused) variable name
  MBS-9847: Show event collections in reverse (#747)
  MBS-9906: Fix link to profile from preferences saved dialog (#746)
  Test for MBS-9748 (Add Ko-fi patronage relationship)
  MBS-9748: URL Cleanup for Ko-fi
  (MBS-9748) Create Ko-fi URL Entity
  MBS-9900:  Add links to request instrument/genre (#752)
  Test cases for MBS-9873
  MBS-9873: Autoselect Support for Discogs Compositions
  MBS-9901: Change Dailymotion cleanup to https (#749)
  MBS-9844: Handle cancel button for apps form
  Require every single field in application form
  Forward required parameter to input/select tag
  Freeze Type field in Application editing form
  Allow freezing selected value in form row
  Add component for form hidden field
  Move ApplicationForm to static/scripts (MBS-9817)
  Factorize code for register/edit application forms
  Move/rename RegisterForm file to ApplicationForm
  Allow disabling FormRowSelect
  Use bracketed function in Footer
  Run eslint on types.js and Footer.js
  Replace React items array with Fragment in Footer
  Factorize type definitions for application forms
  Rewrite Edit Application form to React/JSX
  Rewrite Revoke Application Access page to React/JSX
  Rewrite the Remove Application page to React/JSX
  Factorize code for Confirmation dialog pages
  MBS-9883: More clear sidebar name for Wikidata rels (#738)
  Sync editorMayApprove JS with server logic
  Move edit type constants to new file
  MBS-9892: Allow Amazon Australia links (#743)
  MBS-9889: Fix collection link on sidebar (#740)
  Bump Flow to 0.86.0
  Adding the first batch of genre requests
  Grammatical errors removed.
  Bump Flow to 0.85.0
  Update URL.pm
  MBS-9859: Amazon cover art stats exclude https
  Rewrite the Lost Username page to React/JSX
  Rewrite the Lost Password page to React/JSX
  Convert form row for email field to React

+++ lib/MusicBrainz/Server/Controller/Account.pm
@@ -4,6 +4,8 @@ BEGIN { extends 'MusicBrainz::Server::Controller' }
 
 use namespace::autoclean;
 use Digest::SHA qw(sha1_base64);
+use JSON;
+use List::MoreUtils qw( uniq );
 use MusicBrainz::Server::ControllerUtils::JSON qw( serialize_pager );
 use MusicBrainz::Server::Translation qw( l );
 use MusicBrainz::Server::Validation qw( encode_entities is_positive_integer );
@@ -204,7 +206,14 @@ sub lost_password : Path('/lost-password') ForbiddenOnSlaves
         }
     }
 
-    $c->stash->{form} = $form;
+    $c->stash(
+        current_view => 'Node',
+        component_path => 'account/LostPassword',
+        component_props => {
+            form => $form,
+        },
+    );
+    $c->detach;
 }
 
 sub reset_password : Path('/reset-password') ForbiddenOnSlaves DenyWhenReadonly
@@ -321,7 +330,14 @@ sub lost_username : Path('/lost-username') ForbiddenOnSlaves
         }
     }
 
-    $c->stash->{form} = $form;
+    $c->stash(
+        current_view => 'Node',
+        component_path => 'account/LostUsername',
+        component_props => {
+            form => $form,
+        },
+    );
+    $c->detach;
 }
 
 =head2 edit
@@ -472,6 +488,22 @@ sub preferences : Path('/account/preferences') RequireAuth DenyWhenReadonly
         $c->persist_user();
 
         $c->response->redirect($c->uri_for_action('/account/preferences', { ok => 1 }));
+        $c->detach;
+    } else {
+        $c->stash(
+            current_view => 'Node',
+            component_path => 'account/Preferences',
+            component_props => {
+                form => $form,
+                timezone_options => {
+                    grouped => JSON::false,
+                    options => [ map { {
+                        value => $_,
+                        label => $_
+                      } } uniq values @{ $form->options_timezone } ],
+                },
+            },
+        );
         $c->detach;
     }
 }
@@ -682,10 +714,24 @@ sub revoke_application_access : Path('/account/applications/revoke-access') Args
 
     my $form = $c->form( form => 'SubmitCancel' );
     if ($c->form_posted && $form->submitted_and_valid($c->req->params)) {
-        $c->model('MB')->with_transaction(sub {
-            $c->model('EditorOAuthToken')->revoke_access($c->user->id, $application_id, $scope);
-        });
-        $c->response->redirect($c->uri_for_action('/account/applications'));
+        if ($form->field('cancel')->input) {
+            $c->response->redirect($c->uri_for_action('/account/applications'));
+            $c->detach;
+        } else {
+            $c->model('MB')->with_transaction(sub {
+                $c->model('EditorOAuthToken')->revoke_access($c->user->id, $application_id, $scope);
+            });
+            $c->response->redirect($c->uri_for_action('/account/applications'));
+            $c->detach;
+        }
+    } else {
+        $c->stash(
+            current_view => 'Node',
+            component_path => 'account/applications/RevokeAccess',
+            component_props => {
+                form => $form,
+            },
+        );
         $c->detach;
     }
 }
@@ -737,6 +783,16 @@ sub edit_application : Path('/account/applications/edit') Args(1) RequireAuth Re
         });
         $c->response->redirect($c->uri_for_action('/account/applications'));
         $c->detach;
+    } else {
+        $form->field('oauth_type')->value($application->oauth_type),
+        $c->stash(
+            current_view => 'Node',
+            component_path => 'account/applications/Edit',
+            component_props => {
+                form => $form,
+            },
+        );
+        $c->detach;
     }
 }
 
@@ -750,10 +806,24 @@ sub remove_application : Path('/account/applications/remove') Args(1) RequireAut
 
     my $form = $c->form( form => 'SubmitCancel' );
     if ($c->form_posted && $form->submitted_and_valid($c->req->params)) {
-        $c->model('MB')->with_transaction(sub {
-            $c->model('Application')->delete($application->id);
-        });
-        $c->response->redirect($c->uri_for_action('/account/applications'));
+        if ($form->field('cancel')->input) {
+            $c->response->redirect($c->uri_for_action('/account/applications'));
+            $c->detach;
+        } else {
+            $c->model('MB')->with_transaction(sub {
+                $c->model('Application')->delete($application->id);
+            });
+            $c->response->redirect($c->uri_for_action('/account/applications'));
+            $c->detach;
+        }
+    } else {
+        $c->stash(
+            current_view => 'Node',
+            component_path => 'account/applications/Remove',
+            component_props => {
+                form => $form,
+            },
+        );
         $c->detach;
     }
 }
